pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "autotune"
version = "0.1.0"
description = "Automatic local-LLM inference configuration recommender"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "psutil>=5.9",
    "rich>=13.0",
    "py-cpuinfo>=9.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "httpx>=0.27",
]

[project.scripts]
autotune = "autotune.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["autotune"]
